fix(tsc): improve regexp performance for global type removal
The issue with the removal of global types was the RegExp, for
certain files written like `.tsbuildinfo` which is a single line
long contents file it would scan the whole line scanning for the
characters.

By specifying `m` flag it causes ^ and $ to match the begin/end
of each line (not only begin/end of string).

This removes the need for the check of the filename to do the
replacement check.